requirements.txt
streamlit == 1.36.0
pandas == 2.2.2
numpy == 1.26.4
pyrebase4 == 4.8.0
ultralytics == 8.2.53
opencv-python==4.10.0.84
psutil